% Generated by roxygen2 (4.0.1): do not edit by hand
\name{check_func}
\alias{check_func}
\title{Linear Programming - Check function}
\usage{
check_func(r, tau)
}
\arguments{
\item{r}{A vector to feed into the check function}

\item{tau}{Tau values to be used in the check function}
}
\description{
\code{check_func} Check function
}
\examples{
r <- seq(-1,1,length.out=1e4)
tau <- 0.5
check_func(r,tau)
}

