% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{engel}
\alias{engel}
\title{Engel's Law - Engel Food Expenditures Data from the quantreg package for R}
\format{A dataset containing 235 observations on 2 variables}
\source{
This dataset was used in Koenker and Bassett (1982) and obtained from the quantreg package for R.  Citations:  Koenker, R. and Bassett, G (1982) Robust Tests of Heteroscedasticity based on Regression Quantiles; Econometrica 50, 43-61.
\url{http://CRAN.R-project.org/package=quantreg}
}
\arguments{
\item{income}{Annual household income (Belgian francs)}

\item{foodexp}{Annual household food expenditure (Belgian francs)}
}
\description{
Data on income and food expenditure for 235 working class households in 1857 Belgium.
}
\keyword{datasets}

