\name{example.alpha}
\alias{example.alpha}
\title{Gradient Descent Algorithm - Plots Depicting Gradient Descent Results in Example 1 Using Different Choices for the Step Size}
\usage{
  example.alpha(alpha)
}
\arguments{
  \item{alpha}{Step-size for the gradient descent
  algorithm.}
}
\description{
  \code{example.alpha} Plots the function values for
  gradient descent in Example 1 (without the intercept)
  given a particular value for alpha.
}
\examples{
example.alpha(0.01)
example.alpha(0.12)
}
\author{
  Jocelyn T. Chi
}

