% Generated by roxygen2 (4.0.1): do not edit by hand
\name{generate_data}
\alias{generate_data}
\title{Linear Programming - Generate simulated data for LAD regression}
\usage{
generate_data(n = 100, p = 1, sigma = 1, seed = 12345)
}
\arguments{
\item{n}{Number of data points}

\item{p}{Number of total covariates}

\item{sigma}{Standard deviance}

\item{seed}{Random seed}
}
\description{
\code{generate_data} Function for generating simulated data for LAD regression
}
\examples{
generate_data(n=1000,p=100)
}
\author{
Jocelyn T. Chi
}

