% Generated by roxygen2 (4.0.1): do not edit by hand
\name{generate_sparse_data}
\alias{generate_sparse_data}
\title{Linear Programming - Generate simulated data for sparse quantile regression}
\usage{
generate_sparse_data(n = 100, p = 20, k = 3, seed = 12345)
}
\arguments{
\item{n}{Number of data points}

\item{p}{Number of total covariates}

\item{k}{Number of true covariates}

\item{seed}{Random seed}
}
\description{
\code{generate_sparse_data} Function for generating simulated data for sparse quantile regression
}
\examples{
data <- generate_sparse_data(n=100,p=20,k=3)
}
\author{
Jocelyn T. Chi
}

