\docType{package}
\name{gettingtothebottom}
\alias{gettingtothebottom}
\alias{gettingtothebottom-package}
\title{gettingtothebottom}
\description{
  Getting to the Bottom is a companion package for the
  "Getting to the Bottom" optimization methods series at
  Statisticisviews.com.  The package contains data and code
  to reproduce the examples in the articles.
}
\author{
  Jocelyn T. Chi
}

