% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ladlp}
\alias{ladlp}
\title{Linear Programming - Least absolute deviations (LAD) regression}
\usage{
ladlp(formula)
}
\arguments{
\item{formula}{Formula for LAD regression (e.g., y~X)}
}
\description{
\code{ladlp} Function for performing least absolute deviations using linear programming
}
\examples{
# Simulate data for LAD regression
m <- 1000 # number of observations
n <- 100 # number of variables
reg <- generate_data(m,n)
X <- reg$X
y <- reg$y

# Obtain LAD solution (not run)
# sol_lad <- ladlp(y~X)
# sol_lad
}
\author{
Jocelyn T. Chi
}

