% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make_noise}
\alias{make_noise}
\title{Linear Programming - Generate matrix of noisy data for sparse quantile regression}
\usage{
make_noise(y, X, p, seed = 12345)
}
\arguments{
\item{y}{Response vector}

\item{X}{True predictor}

\item{p}{Number of desired columns of noise}

\item{seed}{(optional) Random seed (Default seed=12345)}
}
\description{
\code{make_noise} Function for generating matrix of noisy data for sparse quantile regression
}
\examples{
y <- engel$foodexp
X <- engel$income
noisydata <- make_noise(y,X,10)
head(noisydata)
}
\author{
Jocelyn T. Chi
}

