\name{makeLambdaseq}
\alias{makeLambdaseq}
\title{MM Algorithm - Function for making sequence of lambdas for solution paths}
\usage{
  makeLambdaseq(L, lambdaseq_length)
}
\arguments{
  \item{L}{Initial lambda value}

  \item{lambdaseq_length}{Desired length of lambda
  sequence}
}
\description{
  \code{makeLambdaseq} Function for making sequence of
  lambdas for solution paths given starting lambda value
  and desired sequence length
}
\examples{
makeLambdaseq(11,20)
}
\author{
  Jocelyn T. Chi
}

