% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{nutrition}
\alias{nutrition}
\title{The Diet Problem: "Daily Allowances of Nutrients for a Moderately Active Man (weighing 154 pounds)" from George Stigler's 1945 paper on "The Cost of Subsistence"}
\format{A vector of length 9}
\source{
George J. Stigler, "The Cost of Subsistence", Journal of Farm Economics, Vol. 27, No. 2 (May, 1945), pp. 303-314.
\url{http://www.jstor.org/stable/1231810}
}
\arguments{
\item{Calories}{Calories (in kilocalories)}

\item{Protein}{Protein (in grams)}

\item{Calcium}{Calcium (in grams)}

\item{Iron}{Iron (in milligrams)}

\item{Vitamin.A}{Vitamin A (in 1000 Interntional Units)}

\item{Thiamine}{Thiamine (in milligrams)}

\item{Riboflavin}{Riboflavin (in milligrams)}

\item{Niacin}{Niacin (in milligrams)}

\item{Ascorbic.Acid}{Ascorbin Acid (in milligrams)}
}
\description{
A vector describing 1943 dietary requirements for a "moderately active" man of 154 pounds.  Obtained from Table 1 of George Stigler's 1945 paper on "The Cost of Subsistence".
}
\keyword{datasets}

