% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot_check}
\alias{plot_check}
\title{Linear Programming - Plot Check Function}
\usage{
plot_check(r, taus = taus)
}
\arguments{
\item{r}{A vector to feed into the check function}

\item{taus}{A vector of tau values to be used in the check function}
}
\description{
\code{plot_check} Plot Check Function
}
\examples{
r <- seq(-1,1,length.out=1e4)
taus <- c(0.25,0.5,0.75)
plot_check(r,taus)
}

