% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_softhreshold}
\alias{plot_softhreshold}
\title{MM Algorithm - Plot the Softhreshold Function}
\usage{
plot_softhreshold(from = -5, to = 5, lambda = 3)
}
\arguments{
\item{from}{The starting value of the sequence of inputs into the function}

\item{to}{The ending value of the sequence of inputs into the function}

\item{lambda}{Lambda value for softhreshold function}
}
\description{
\code{plot_softhreshold} Function for plotting the softhreshold function
}
\examples{
plot_softhreshold(-5,5,3)
}
\author{
Jocelyn T. Chi
}

