% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{stigler}
\alias{stigler}
\title{The Diet Problem: "Nutritive Values of Common Foods per Dollar of Expenditure, August 15, 1944", from George Stigler's 1945 paper on "The Cost of Subsistence"}
\format{A data frame with 10 rows and 14 columns}
\source{
George J. Stigler, "The Cost of Subsistence", Journal of Farm Economics, Vol. 27, No. 2 (May, 1945), pp. 303-314.
\url{http://www.jstor.org/stable/1231810}
}
\arguments{
\item{Calories}{Calories (in kilocalories) per dollar}

\item{Protein}{Protein (in grams) per dollar}

\item{Calcium}{Calcium (in grams) per dollar}

\item{Iron}{Iron (in milligrams) per dollar}

\item{Vitamin.A}{Vitamin A (in 1000 Interntional Units) per dollar}

\item{Thiamine}{Thiamine (in milligrams) per dollar}

\item{Riboflavin}{Riboflavin (in milligrams) per dollar}

\item{Niacin}{Niacin (in milligrams) per dollar}

\item{Ascorbic.Acid}{Ascorbin Acid (in milligrams) per dollar}
}
\description{
A dataset of 10 rows and 14 columns describing 1944 prices and nutritional data for 14 food commodities.  Obtained from George Stigler's 1945 paper on "The Cost of Subsistence".
}
\keyword{datasets}

