% Generated by roxygen2 (4.0.1): do not edit by hand
\name{softhreshold}
\alias{softhreshold}
\title{MM Algorithm - Softhreshold Function}
\usage{
softhreshold(x, lambda)
}
\arguments{
\item{x}{Vector of values to be softhresholded}

\item{lambda}{Softhreshold parameter}
}
\description{
\code{softhreshold} Function for computing the softhreshold
}
\examples{
x <- seq(-10,10,1)
softhreshold(x,lambda=3)
}
\author{
Jocelyn T. Chi
}

